collect-server/requirements.txt
requests
kafka-python
pymongo